requirements.txt
cycler==0.11.0
fonttools==4.38.0
importlib-metadata==6.7.0
kiwisolver==1.4.5
matplotlib==3.5.3
munch==4.0.0
numpy==1.21.6
opencv-python==4.9.0.80
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
scikit-learn==0.19.2
scipy==1.7.3
six==1.16.0
toml==0.10.2
torch==1.7.1
torchvision==0.8.2
tqdm==4.66.1
typing_extensions==4.7.1
zipp==3.15.0
